project-009/my_daily_todos_app/pyproject.toml
[tool.poetry]
name = "my-daily-todos"
version = "0.1.6"
description = "A Daily Todo App with Multi-Users Admin"
authors = ["Syed Muhammad Hussain <hussain.sm014@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlmodel = "^0.0.16"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg = {extras = ["binary"], version = "^3.1.18"}
pytest = "^8.1.1"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
